% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batches.R
\name{.run_batch_3}
\alias{.run_batch_3}
\title{Run multivaritate episode pipeline in batches}
\usage{
.run_batch_3(
  connection,
  ids_subset,
  study_variables = study_variables,
  sql_explosion = sql_explosion,
  sql_combine = sql_combine,
  sql_mergestatus = sql_mergestatus,
  data_type_col = data_type_col
)
}
\arguments{
\item{connection}{A database connection object}

\item{ids_subset}{A vector of person_ids to process in this batch}

\item{study_variables}{A list of study variables}

\item{sql_explosion}{The SQL file for exploding spells}

\item{sql_combine}{The SQL file for combining daily values}

\item{sql_mergestatus}{The SQL file for merging status intervals}

\item{data_type_col}{The column name for data types}
}
\description{
Run multivaritate episode pipeline in batches
}
\keyword{internal}
