% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episode_label}
\alias{episode_label}
\title{Label events meeting episode criteria}
\usage{
episode_label(data, group, IEI, intervals = "Inter_Event_Intervals",
  tag = "Episode")
}
\arguments{
\item{data}{A data frame containing event times along with optional descriptive information.}

\item{group}{A character vector containing the name(s) of one or more grouping factor combinations.
e.g. \code{c("Subject", "Session")}.}

\item{IEI}{A minimum inter-event interval criterion value.}

\item{intervals}{The name of the inter-event interval vector}

\item{tag}{A string identifying episode type. "Episode" is the is default.}
}
\value{
A data frame (tbl_data class; see
\href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
consisting of input data and 4 columns of episode information:
\itemize{
\item Episode_Interval_Criterion: A logical vector corresponding to inter_event_interval runs meeting
the IEI criterion.
\item Episode_Events: An integer vector containing to the number of events within episode.
\item Within_Episode_Order: An integer vector containing the order of events within each episode.
\item Episode_Order: An integer vector containing the order of episodes within each factor combination
}
}
\description{
Identify events conforming to a specified inter-event intervals, the number of
events per episode. Provide order of both events and episodes.
}
\seealso{
\code{\link{episode_summarize}}.
}
