% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episode_order}
\alias{episode_order}
\title{Number vector of repeating elements}
\usage{
episode_order(data, group, episode_crit = "Episode_Crit")
}
\description{
sequence of repeating elmemns where x is data frame vector of episode runs. This operation is expensive with larger opperation dplyr impmentation
}

