% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episode_summarize}
\alias{episode_summarize}
\title{Generate summary of episode parameters}
\usage{
episode_summarize(data, group, IEI, EPE, time = "Time")
}
\arguments{
\item{data}{A data frame containing event and inter-event interval times along with optional
descriptive information.}

\item{group}{A character vector containing the name(s) of one or more grouping factor combinations.
e.g.  \code{c(“Subject”, “Session”)}.}

\item{IEI}{A minimum Inter-Event Inteval criterion value.}

\item{EPE}{A minimum Events Per Episode criterion value.}

\item{time}{The name of the event time vector (as a characther string).}
}
\value{
A data frame (tbl_data class; see
 \href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
 of pisode parmaters. Each row corresponds to an episodes as defined by IEI and EPE episode criteria.
Addtional vectors contain the following episode informaiton:
\itemize{
\item Episode_Events: The number of events per episode.
\item Episode Starts: Episode start time.
\item Episode_End: Episode end time.
\item Event_Duration: Episode duration
\item Event_Duration: Episode rate.
}
}
\description{
Generate a summary of episode paramters from event time data according
to minimum inter-event interval duration and minimum number of events per episode criteria.
}
\examples{
data(lickometer)
episode_summarize(lickometer, group = c("Subject", "Session"), IEI= 12, EPE = 10 ,time ="Time")
}

