% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episode_summarize2}
\alias{episode_summarize2}
\title{Generate summary of episode parameters (for development)}
\usage{
episode_summarize2(data, group, summarize_by, time = "Time")
}
\arguments{
\item{data}{A data frame containing event times along with additional descriptive information}

\item{group}{A character vector of the name(s) of one or more factor combinations for}

\item{summarize_by}{A vector labling events to summarize}

\item{time}{Name of event time vector}
}
\value{
A data frame (tbl_data class; see
\href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
of episodes and episode parmaters. Rows corespond to episodes.
Episode_Events: An integer vector corresponding to the number of events per episode.
Episode_End: An integer vector of episode end times.
Event_Duration: Episdoe duration.
}
\description{
Create data frame summary of episode paramters from labled data.
}
\examples{
data(lickometer)
episode_summarize(lickometer, group = c("Subject", "Session"), IEI= 12, EPE = 10 ,time ="Time")
}

