% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{episoder}
\alias{episoder}
\alias{episoder-package}
\title{episoder: A package for episode and event interval analysis}
\description{
A common approach to analyzing high-resolution behavioural event data, eg. lickometer or
feeder recordings, is by grouping individual events into episodes  according to inter-event-interval
distributions. The episoder package provides several functions for obtaining inter-event-interval and
episode information form large, event time-series datasets.
}
\details{
The episoder futnions \code{\link{event_intervals}}, \code{\link{event_episodes}},
while \code{\link{episode_summarize}},
}
\section{See Also}{

For further infrmoanton see viggnet("episodes") for further informaiton
}

