% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episoder.summarize}
\alias{episoder.summarize}
\title{Create data frame summary episode paramters}
\usage{
episoder.summarize(df, group, IEI, EPE)
}
\arguments{
\item{df}{A data frame}

\item{group}{A characther vector of unique data frame vector factor combinations for each c("Session", "Subject")}

\item{IEI}{A value of time less than or equal the minimum inter-event interval duration criterion for an episode}

\item{EPE}{A value specifiing maximum Event}
}
\description{
Create data frame summary of episode paramters from event time data. Includes summary
}

