% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episoder_identify}
\alias{episoder_identify}
\title{Identify episodes by minimum inter-episode duration}
\usage{
episoder_identify(df, IEI, intervals = "Inter_Event_Intervals")
}
\arguments{
\item{df}{A data frame containing episode data}

\item{IEI}{a value of time less than or equal the minimum inter-event interval duration criterion for an episode}

\item{name}{of data frame vector containg interepisode}
}
\value{
a data frame or tible of the same lengh as the input data frame.
}
\description{
Identify episodes conforming to a minimum interepisode duration criterion
}

