% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episoder_intervals}
\alias{episoder_intervals}
\title{Calcualte inter-event intervals}
\usage{
episoder_intervals(df, group, time = "Time")
}
\arguments{
\item{df}{A data frame}

\item{group}{A characther vector of unique data frame vector factor combinations for each c("Session", "Subject")}

\item{time}{The name of data frame vector containg event times}
}
\value{
New vector "Inter_Episode_Intervals" consiting inter episode intervals appended to input existing df
}
\description{
Transform a vector of event time data into inter-eventduraton interval according to specified factor. The final
inter-event value for each factor level is NA.
}
\seealso{
\code{\link{episoder_identify}} for finding episodes.
}

