% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episoder_order}
\alias{episoder_order}
\title{Number vector of repeating ellements}
\usage{
episoder_order(df, group, episode_crit = "Episode_crit")
}
\description{
sequence of repeating elmemns where x is data frame vector of episode runs. This operation is expensive with larger opperation dplyr impmentation
}

