% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event_bins}
\alias{event_bins}
\title{Add event time bins}
\usage{
event_bins(data, time, bins, samples_per_session, samples_per_bin)
}
\arguments{
\item{data}{A data frame containing event times along with additional descriptive information}

\item{time}{The name of the event time column}

\item{bins}{A vector containg bin lables}

\item{samples_per_session}{the maximum time units units, or samples in a session}

\item{samples_per_bin}{number of samples, or time units per bin}
}
\description{
add event time bins
}
\examples{
data(lick_data)
event_bins(licks_df, time = "Time",
samples_per_session = 50 *60 * 60 * 23,
samples_per_bin = 50 *60 * 60, bins = c(0:22))
}

