% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event_survival}
\alias{event_survival}
\title{Prepare events for survival analysis}
\usage{
event_survival(data, IEI_min, IEI_max, time_start, time_end,
  group = c("Treamtment"))
}
\arguments{
\item{data}{A data frame containing a vector of event runs}

\item{group}{A character vector containing the name(s) of one or more grouping factor combinations.}

\item{IEI_range}{Lower cut off inter-event interval duration}

\item{time_range}{Upper cut off inter-event interval duration}

\item{time_cutoff}{Upper cut for time}
}
\value{
A data frame (tbl_data class; see
\href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
of episodes and episode parameters. Rows correspond to episodes.
Episode_Events: An integer vector corresponding to the number of events per episode.
Episode_End: An integer vector of episode end times.
Event_Duration: Episode duration.
}
\description{
Order and summat events for survival analsys
}
