% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_episode.R
\name{get_episode}
\alias{get_episode}
\title{Get episodes of variable growth and decline episodes}
\usage{
get_episode(
  data = EpisodeR::vdem,
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  variable = "v2xca_academ"
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set.}

\item{start_incl}{What is the minimum annual change of a variable necessary to trigger an episode? This is the absolute value of the first difference
in the variable required for the onset of either a decline or growth episode}

\item{cum_incl}{What is the minimum amount of total change on the variable necessary to constitute a manifest episode?
A potential episode might be a period involving any amount of changes over a period following an annual change equal
to the start inclusion (e.g. 0.01). To identify substantial changes, we set a cumulative inclusion threshold.
This is the absolute value of the total amount of change needed on the EDI to be considered manifest.}

\item{year_turn}{What is the amount of annual change in the opposite direction to trigger the termination of an episode?
An episode may end when the case suddenly moves in the opposite direction.}

\item{variable}{What is the variable the dataset looks for}
}
\value{
A data frame specifying episodes of growth and decline of a specific variable and their outcomes in the most recent V-Dem data set.
For further details and explanations on episodes and outcomes please check the respective journal article:
Lott, Lars (2023). Academic Freedom Growth and Decline Episodes. Higher Education. DOI: 10.1007/s10734-023-01156-z
}
\description{
\code{get_episode} returns a data.frame identifying decline and growth episodes
of a variable in which this variable systematically grow or decline in respective country-years. This function controls for overlapping
confidence /uncertainty intervals as suggested by Pelke, Lars & Aurel Croissant (2021). Conceptualizing and Measuring Autocratization Episodes.
Swiss Political Science Review, 27 (2), 434-448. doi:10.1111/spsr.12437
}
\details{
\emph{increase_episode} is an umbrella term for any movement towards more academic freedom that is statistically significant, i.e. outside of the uncertainty interval
\emph{decline_episode} is an umbrella term for any movement away from academic freedom that is statistically significant, i.e. outside of the uncertainty interval
}
