% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licks_df.R
\docType{data}
\name{licks_df}
\alias{licks_df}
\title{Lickometer data for 8 rats.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1278038 rows and 7 columns.}
\usage{
licks_df
}
\description{
Lickometer data collected from 8 rats over 2 session in 2 treatment condtions.
}
\details{
\describe{
  \item{Subject}{Subject ID}
  \item{Session}{Session number}
  \item{Time}{Licks recorded at a 20 ms sampling resolution}
  \item{Treatment}{Treatment group}
  ...
}
}
\keyword{datasets}

