% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_episodes_pipeline_7.R
\name{multivariate_episodes_pipeline_7}
\alias{multivariate_episodes_pipeline_7}
\title{Run the multivariate episodes pipeline, version 7.}
\usage{
multivariate_episodes_pipeline_7(connection, input_directory, output_directory)
}
\arguments{
\item{connection}{A connection to a database where the multivariate episodes pipeline will be executed.}

\item{input_directory}{Directory of the univariate episodes parquet files.}

\item{output_directory}{Directory where the output parquet files will be saved.}
}
\description{
Run the multivariate episodes pipeline, version 7.
}
