% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_episodes.R
\name{plot_all_episodes}
\alias{plot_all_episodes}
\title{Plotting the global number/share of countries undergoing growth and decline episodes}
\usage{
plot_all_episodes(
  abs = T,
  years = c(1900, 2023),
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  variable = "v2xca_academ"
)
}
\arguments{
\item{abs}{Logical value: if TRUE, the absolute number of countries in an episode for each year is plotted.
If FALSE, the share of countries in percentage undergoing growth or decline is plotted.}

\item{years}{Vector with two numeric values indicating the minimum and maximum year to be plotted.}

\item{start_incl}{What is the minimum annual change of a variable necessary to trigger an episode? This is the absolute value of the first difference
in the variable required for the onset of either a decline or growth episode}

\item{cum_incl}{What is the minimum amount of total change on the variable necessary to constitute a manifest episode?
A potential episode might be a period involving any amount of changes over a period following an annual change equal
to the start inclusion (e.g. 0.01). To identify substantial changes, we set a cumulative inclusion threshold.
This is the absolute value of the total amount of change needed on the index variable to be considered manifest.}

\item{year_turn}{What is the amount of annual change in the opposite direction to trigger the termination of an episode?
An episode may end when the case suddenly moves in the opposite direction.

The terms @param abs, @param years, @param start_incl, @param cum_incl, and @param year_turn as well as the descriptions of these terms in the function are adapted
from the ERT-package available at https://github.com/vdeminstitute/ERT. In the function below, all parts of code that was copied and adapted from the ERT package are tagged.
The original ERT package enable users to set additional parameters to customize their definitions of what constitutes an episode of change. These additional parameters are
the tolerance parameter, and the cum_turn parameter. In this package, episodes are considered as an episode as long as there is continued increase/decline,
while allowing up to 4 years of temporary stagnation. This period of 4 years cannot be set to another value of temporary stagnation (compare the more flexible ERT package).}

\item{variable}{What is the index variable the dataset looks for. By default, the v2xca_academ variable is loaded.}
}
\value{
The output of this function is a ggplot() object with the number/share of autocratization episodes per year.
}
\description{
\code{plot_all_episodes} plots the global number or share of countries undergoing
growth and decline episodes in a specific variable in a selected time frame.
}
\details{
This function is a wrapper for ggplot() and produces a plot that shows
growth and decline episodes over time for a specific variable
The function calls the EpisodeR:get_episode_wo_CI() function to identify episodes,
but without controlling for overlapping confidence intervals
}
