% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_episodes.R
\name{plot_episodes}
\alias{plot_episodes}
\title{Plot episodes of variable growth and decline episodes for a specific country}
\usage{
plot_episodes(
  years = c(1900, 2021),
  country = c("Sweden"),
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  variable = "v2xca_academ"
)
}
\arguments{
\item{years}{Vector with two numeric values indicating the minimum and maximum year to be plotted.}

\item{country}{Character vector containing the country for which episodes should be shown. Only entries from the
country_name column in the V-Dem data set are accepted.}

\item{start_incl}{What is the minimum annual change of a variable necessary to trigger an episode? This is the absolute value of the first difference
in the variable required for the onset of either a decline or growth episode}

\item{cum_incl}{What is the minimum amount of total change on the variable necessary to constitute a manifest episode?
A potential episode might be a period involving any amount of changes over a period following an annual change equal
to the start inclusion (e.g. 0.01). To identify substantial changes, we set a cumulative inclusion threshold.
This is the absolute value of the total amount of change needed on the EDI to be considered manifest.}

\item{year_turn}{What is the amount of annual change in the opposite direction to trigger the termination of an episode?
An episode may end when the case suddenly moves in the opposite direction.}

\item{variable}{What is the variable the dataset looks for}

\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set.}
}
\value{
The output of this function is a \code{\link[=ggplot2:ggplot]{ggplot2:ggplot()}} object with the number/share of autocratization episodes per year.
}
\description{
\code{plot_episodes} plots growth and decline episodes in a specific variable
over time for a selected country and a selected time frame.
}
\details{
This function is a wrapper for \code{\link[=ggplot2:ggplot]{ggplot2:ggplot()}} and produces a plot that shows
growth and decline episodes for a selected country over time.
The function calls the \code{\link[=EpisodeR:get_episode_wo_CI]{EpisodeR:get_episode_wo_CI()}} function to identify episodes, but without controlling for overlapping confidence intervals
}
