% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{renumber}
\alias{renumber}
\title{Renumber a vector of event runs}
\usage{
renumber(data, group, name_in, name_out)
}
\arguments{
\item{data}{A data frame containing a vector of event runs}

\item{group}{A character vector of the name(s) of one or more factor combinations for}

\item{name_in}{Name of the vector to renumber}

\item{name_out}{Name of renumbered vector. If left blank, the name is unchanged}
}
\value{
A data frame (tbl_data class; see
\href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
of episodes and episode parmaters. Rows corespond to episodes.
Episode_Events: An integer vector corresponding to the number of events per episode.
Episode_End: An integer vector of episode end times.
Event_Duration: Episdoe duration.
}
\description{
Squentially number runs of equal values in a vector within a data frame
}
\examples{
data(lickometer)
episode_summarize(lickometer, group = c("Subject", "Session"), IEI= 12, EPE = 10 ,time ="Time")
}

