% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_sequence}
\alias{run_sequence}
\title{Order of runs of equal values in a vector}
\usage{
run_sequence(x)
}
\arguments{
\item{x}{A vector containg runs of equal values}
}
\value{
a vector equal in length to x, consiting an incrementing
integers corespoidng to each run of equal values in x
}
\description{
Generate ordered sequnce of intergers for each run of equal values
in a vector
for number repeating event
}
\seealso{
\code{\link{rle}}
}

