% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_seq}
\alias{seq_seq}
\title{Repeat each element of of vector (vector[i]), vector[i] times}
\usage{
seq_seq(x)
}
\arguments{
\item{x}{A vector integers
#' @seealso \code{\link{sequence}} and \code{\link{seq}}}
}
\description{
sequence for each element the value of that element times. Usefull along with wiht  \code{\link{rle}}
   for number repeating event
}

