% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_bins}
\alias{add_bins}
\title{Add time bins}
\usage{
add_bins(data, time, samples_per_session, samples_per_bin, bin_labels,
  bin_name = "Bins")
}
\arguments{
\item{data}{A data frame containing a vector of event-times along plus optional descriptive information.}

\item{time}{The name of the event-time vector (as a character string).}

\item{samples_per_session}{The total number of samples, or maximum time unit number of a session.}

\item{samples_per_bin}{The number of samples, or time units per bin. Must be a dividend of
samples_per_session.}

\item{bin_labels}{A vector containing bin labels. Length of the vector must corrospond
to the number of bins. If missing, bin labels are added automatically (0:number of bins).}

\item{bin_name}{Name of binning vector. Default name is "Bins"}

\item{samples_per_bin}{The number of samples per bin. Must be a dividend of samples_per_session.}
}
\description{
Add a binning vector for event-time data.
}
\examples{
data(lickometer)
add_bins(lickometer, time = "Time", samples_per_session = 50 *60 * 60 * 23, samples_per_bin = 50 *60 * 60,
bin_labels = 0:22)
}
