% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{episode_summarize}
\alias{episode_summarize}
\title{Generate summary of episode parameters}
\usage{
episode_summarize(data, group, IEI, EPE, intervals = "Inter_Event_Intervals",
  time = "Time")
}
\arguments{
\item{data}{A data frame containing event times along with additional descriptive information}

\item{group}{A character vector of the name(s) of one or more factor combinations for}

\item{IEI}{A minimum inter-interval episode criterion value.}

\item{EPE}{Mximum Events per Episode value}

\item{time}{The name of the event time column}
}
\value{
A data frame (tbl_data class; see
\href{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}{Tibbles})
of episodes and episode parmaters. Rows corespond to episodes.
Episode_Events: An integer vector corresponding to the number of events per episode.
Episode_End: An integer vector of episode end times.
Event_Duration: Episdoe duration.
}
\description{
Create data frame summary of episode paramters from event time data. Includes summary
}
\examples{
data(lickometer)
episode_summarize(lickometer, group = c("Subject", "Session"), IEI= 12, EPE = 10 ,time ="Time")
}

