% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{episoder}
\alias{episoder}
\alias{episoder-package}
\title{episoder: A package for computating time series event data into episode}
\description{
The episoder package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

