% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event_episodes}
\alias{event_episodes}
\title{Lable events minimum inter-episode interval criterion}
\usage{
event_episodes(df, IEI, intervals = "Inter_Event_Intervals")
}
\arguments{
\item{df}{A data frame containing inter-event intervals along with additional descriptive information}

\item{IEI}{An inter-event value less than or equal to desired episode criterion}

\item{intervals}{the name of column containg inter-event}
}
\value{
a data frame or tible of the same lengh as the input data frame along with 3 new columns
Episode_crit: A logical vector correspondign to event runs meeting the IEI criterion
Episode.events
}
\description{
Identify episodes conforming to a minimum interepisode duration criterion
}

