% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_episode.R
\name{get_episode}
\alias{get_episode}
\title{Get episodes of variable growth and decline episodes}
\usage{
get_episode(
  data = EpisodeR::vdem,
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  variable = "v2xca_academ"
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set is loaded.}

\item{start_incl}{What is the minimum annual change of a variable necessary to trigger an episode? This is the absolute value of the first difference
in the variable required for the onset of either a decline or growth episode.}

\item{cum_incl}{What is the minimum amount of total change on the index variable necessary to constitute a growth or decline episode?
A potential episode might be a period involving any amount of changes over a period of four years following an annual change equal
to the start inclusion (e.g. 0.01). To identify substantial changes, we set a cumulative inclusion threshold.
This is the absolute value of the total amount of change needed on the index variable to be considered manifest.}

\item{year_turn}{What is the amount of annual change in the opposite direction to trigger the termination of an episode?
An episode may end when the case suddenly moves in the opposite direction.

The terms @param start_incl, @param cum_incl, and @param year_turn as well as the descriptions of these terms in the function are adapted
from the ERT-package available at https://github.com/vdeminstitute/ERT. In the function below, all parts of code that was copied and adapted from the ERT package are tagged.
The original ERT package enable users to set additional parameters to customize their definitions of what constitutes an episode of change. These additional parameters are
the tolerance parameter, and the cum_turn parameter. In this package, episodes are considered as an episode as long as there is continued increase/decline,
while allowing up to 4 years of temporary stagnation. This period of 4 years cannot be set to another value of temporary stagnation (compare the more flexible ERT package).}

\item{variable}{What is the index variable the dataset looks for. By default, the v2xca_academ variable is loaded.}
}
\value{
A data frame specifying episodes of growth and decline of a specific variable and their outcomes in the most recent V-Dem data set.
For further details and explanations on episodes and outcomes please check the respective journal article:
Lott, Lars (2023). Academic Freedom Growth and Decline Episodes. Higher Education. DOI: 10.1007/s10734-023-01156-z
}
\description{
\code{get_episode} returns a data.frame identifying decline and growth episodes
of an index variable in which this variable systematically grow or decline in respective country-years. This function controls for overlapping
confidence/uncertainty intervals as suggested by Pelke, Lars & Aurel Croissant (2021). Conceptualizing and Measuring Autocratization Episodes.
Swiss Political Science Review, 27 (2), 434-448. doi:10.1111/spsr.12437
}
\details{
\emph{increase_episode} is an umbrella term for any movement towards more academic freedom that is statistically significant, i.e. outside of the uncertainty interval
\emph{decline_episode} is an umbrella term for any movement away from academic freedom that is statistically significant, i.e. outside of the uncertainty interval
}
