% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licks_df.R
\docType{data}
\name{licks_df}
\alias{licks_df}
\title{Lickometer data from}
\format{A data frame with 1278038 rows and n variables:
\describe{
  \item{Subject} {Subject number}
  \item{Session}{Session number of test}
  \item{Time}{Event time recorded at a 20 ms samppling resolution}
  \item{Tr}{Treatmentatment group}
  ...
}}
\usage{
licks_df
}
\description{
A dataset containg lick data from 8 rats over
2 23.5 hour sessions
}
\keyword{datasets}

