% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licks_df.R
\docType{data}
\name{licks_df}
\alias{licks_df}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Subject}{Number of thew}
  \item{Session}{Number of test}
  \item{Time}{Event time recorded at a 20 ms samppling resolution}
  \item{Tr}{Treatmentatment group}#'
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
licks_df
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

