% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_sequence}
\alias{run_sequence}
\title{Sequence of equal value runs in a vector}
\usage{
run_sequence(x)
}
\arguments{
\item{x}{A vector containing runs of equal values.}
}
\value{
A vector, equal in length to x, consisting of integers incrementing for each run
of equal values in x.
}
\description{
Generate an ordered sequence of integers corresponding to each run of equal values
in a vector.
}
\seealso{
\code{\link{rle}},  \code{\link{renumber}}
}
