% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_seq}
\alias{seq_seq}
\title{Generate sequence of vector[i] repeated vector[i] times}
\usage{
seq_seq(x)
}
\arguments{
\item{x}{A vector integers}
}
\description{
sequence considitng  for each element of vector repeated  element times.
Usefull along incombination with for   \code{\link{rle}}
for number repeating event
}
\seealso{
\code{\link{sequence}} and \code{\link{seq}}
}

